Add main.cpp
Add main for testing the program
this main in this stage of development
would compile but not pass the linker
stage as there are no source files to be
linked.

This main should test the core functionalities
of the system and give a loop handler to test
series of action one after the other.